requirements.txt
fastapi
uvicorn
requests
httpx
sqlalchemy
pydantic-settings
bcrypt
alembic